requirements.txt
tensorflow-gpu==1.12.0
numpy==1.14.3
matplotlib==2.2.2
Pillow==5.1.0
pickleshare==0.7.4
scipy==1.1.0
scikit-image==0.13.1
sortedcollections==0.6.1
